% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blue.R
\name{build_data}
\alias{build_data}
\title{build_data}
\usage{
build_data(data, metadata, vers)
}
\arguments{
\item{data}{A dataframe with columns \code{datetime} and \code{temperature}}

\item{metadata}{A list with elements \code{serial}, \code{name}, \code{lat}, \code{long}, \code{unit}, \code{resolution}, and \code{tz}}

\item{vers}{A list with elements \code{vers_logger} and \code{vers_software}}
}
\value{
A dataframe with columns \code{datetime}, \code{temperature}, and \code{unit}
}
\description{
Build the data frame
}
