% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blue.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(data)
}
\arguments{
\item{data}{A dataframe with columns \code{datetime} and \code{temperature}}
}
\value{
A list with elements \code{serial}, \code{name}, \code{lat}, \code{long}, \code{unit}, \code{resolution}, and \code{tz}
}
\description{
Get the metadata from the electric blue file
}
