% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blue.R
\name{get_versions}
\alias{get_versions}
\title{get_versions}
\usage{
get_versions(data)
}
\arguments{
\item{data}{A dataframe with columns \code{datetime} and \code{temperature}}
}
\value{
A list with elements \code{vers_logger} and \code{vers_software}
}
\description{
Get the software and logger versions from the data
}
