% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blue.R
\name{list_class}
\alias{list_class}
\title{list_class}
\usage{
list_class(x)
}
\arguments{
\item{x}{A dataframe with columns \code{datetime}, \code{temperature}, and \code{unit}}
}
\value{
A dataframe with class \code{blue} for eache element of the list
}
\description{
Set the class of the data
}
