% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.blue.R
\name{plot.blue}
\alias{plot.blue}
\title{plot.blue}
\usage{
\method{plot}{blue}(data, ...)
}
\arguments{
\item{data}{A data frame containing the temperature data}

\item{...}{Additional arguments to be passed to the plot function}
}
\value{
A plot of the temperature data over time
}
\description{
A function to plot temperature data in blue from teh raw data
}
