% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blue.R
\name{read_blue}
\alias{read_blue}
\title{read_blue}
\usage{
read_blue(dir, path = NULL, ...)
}
\arguments{
\item{dir}{A character string giving the path to the directory to read}

\item{...}{Additional arguments passed to \code{\link{readr::read_csv}}}

\item{file}{A character string giving the path to the file to read. Defaults to NULL. If both \code{dir} and \code{file} are provided, \code{file} will be used.}
}
\value{
A dataframe with columns \code{datetime}, \code{temperature}, and \code{unit}
}
\description{
Read an electric blue temperature logging file or directory of files
}
\examples{
\dontrun{
data <- read_blue("./raw")
plot(data)
}
}
