% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blue.R
\name{read_blue}
\alias{read_blue}
\title{read_blue}
\usage{
read_blue(path, ...)
}
\arguments{
\item{path}{A character string giving the path to the file or directory to read}

\item{...}{Additional arguments passed to \code{\link{readr::read_csv}}}
}
\value{
A dataframe with columns \code{datetime}, \code{temperature}, and \code{unit}
}
\description{
Read an electric blue temperature logging file or directory of files
}
\examples{
\dontrun{
data <- read_blue("./raw/04BC 0000 1C02 10-20240330 125652.csv")
plot(data)
}
}
